Incorporated review comments: * Updated eventually to consistently func for cleanup failure flag * added kubeconfig flag zero value test case * moved helper funcs to test_helper.go file
